Renamed auth module to home module
