+++ app/src/main/java/com/example/newzz/AppModule.kt
@@ -1,12 +1,12 @@
 package com.example.newzz
 
-import com.example.newzz.authentication.authModule
+import com.example.newzz.home.homeModule
 import com.example.newzz.base.getDataModule
 import org.koin.core.module.Module
 
 fun getModule(): List<Module> {
     return mutableListOf<Module>().apply {
         addAll(getDataModule())
-        add(authModule)
+        add(homeModule)
     }
 }
