--- auth/src/main/java/com/example/newzz/authentication/AuthModule.kt
@@ -1,9 +1,0 @@
-package com.example.newzz.authentication
-
-import com.example.newzz.authentication.ui.splash.SplashViewModel
-import org.koin.androidx.viewmodel.dsl.viewModel
-import org.koin.dsl.module
-
-val authModule = module {
-    viewModel { SplashViewModel() }
-}
