+++ base/build.gradle
@@ -50,7 +50,7 @@ android {
     productFlavors {
         localTest {
             buildConfigField 'String', 'APP_NAME', '"newzz"'
-            buildConfigField 'String', 'BASE_URL', '""'
+            buildConfigField 'String', 'BASE_URL', '"https://newsapi.org/v2/everything"'
             buildConfigField 'String', 'NEWS_API_KEY', '"dbf1909199b5492db8953081776fbe1a"'
             buildConfigField 'String', 'VERSION_REALM_SCHEMA', '"0"'
             dimension "default"
@@ -58,7 +58,7 @@ android {
 
         dev {
             buildConfigField 'String', 'APP_NAME', '"newzz"'
-            buildConfigField 'String', 'BASE_URL', '""'
+            buildConfigField 'String', 'BASE_URL', '"https://newsapi.org/v2/everything"'
             buildConfigField 'String', 'NEWS_API_KEY', '"dbf1909199b5492db8953081776fbe1a"'
             buildConfigField 'String', 'VERSION_REALM_SCHEMA', '"0"'
             dimension "default"
@@ -66,7 +66,7 @@ android {
 
         production {
             buildConfigField 'String', 'APP_NAME', '"newzz"'
-            buildConfigField 'String', 'BASE_URL', '""'
+            buildConfigField 'String', 'BASE_URL', '"https://newsapi.org/v2/everything"'
             buildConfigField 'String', 'NEWS_API_KEY', '"dbf1909199b5492db8953081776fbe1a"'
             buildConfigField 'String', 'VERSION_REALM_SCHEMA', '"0"'
             dimension "default"
