+++ home/src/androidTest/java/com/example/newzz/home/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.newzz.authentication
+package com.example.newzz.home
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -17,6 +17,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.newzz.authentication.test", appContext.packageName)
+        assertEquals("com.example.newzz.home.test", appContext.packageName)
     }
 }
