+++ home/src/main/java/com/example/newzz/home/HomeModule.kt
@@ -1,0 +1,9 @@
+package com.example.newzz.home
+
+import com.example.newzz.home.ui.splash.SplashViewModel
+import org.koin.androidx.viewmodel.dsl.viewModel
+import org.koin.dsl.module
+
+val homeModule = module {
+    viewModel { SplashViewModel() }
+}
