+++ home/src/main/res/layout/fragment_splash.xml
