+++ home/src/main/res/navigation/auth_nav_graph.xml
@@ -6,7 +6,7 @@
     app:startDestination="@id/splash">
 
     <fragment android:id="@+id/splash"
-        android:name="com.example.newzz.authentication.ui.splash.SplashFragment"
+        android:name="com.example.newzz.home.ui.splash.SplashFragment"
         tools:layout="@layout/fragment_splash">
 <!--        <action android:id="@+id/action_splash_to_onboard"-->
 <!--            app:destination="@id/on_board"/>-->
