+++ home/src/main/res/values-vi/strings.xml
