+++ home/src/main/res/values/strings.xml
